Bug fix on integer's for mincut

+++ dbmodel/ws/fct/ws_gw_fct_mincut_inv_flowtrace.sql
@@ -58,7 +58,7 @@ BEGIN
 			*/
 		
 			query_text:= 'SELECT * FROM pgr_dijkstra( 
-				''SELECT v_edit_arc.arc_id::integer as id, node_1::integer as source, node_2::integer as target, 
+				''SELECT v_edit_arc.arc_id::integer as id, node_1::int8 as source, node_2::int8 as target, 
 				(case when closed=true then -1 else 1 end) as cost,
 				(case when closed=true then -1 else 1 end) as reverse_cost
 				FROM SCHEMA_NAME.v_edit_arc 
@@ -77,7 +77,7 @@ BEGIN
 						OR (node_2 IN (SELECT node_id FROM SCHEMA_NAME.anl_mincut_result_valve WHERE closed=TRUE AND proposed IS NOT TRUE AND result_id='||result_id_arg||'))
 
 					)a ON a.arc_id=v_edit_arc.arc_id
-				WHERE node_1 is not null and node_2 is not null'','||rec_valve.node_id||'::integer, '||rec_tank.node_id||'::integer)';
+				WHERE node_1 is not null and node_2 is not null'','||rec_valve.node_id||'::int8, '||rec_tank.node_id||'::int8)';
 
 			IF query_text IS NOT NULL THEN	
 				IF (select value::boolean from config_param_system where parameter='om_mincut_valve2tank_traceability') IS TRUE THEN
